pagination and others completed

+++ network/views.py
@@ -8,15 +8,46 @@ import json
 from .models import User, Post, Relationships
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
+from django.core.paginator import Paginator
 
 
 def index(request):
     return render(request, "network/index.html")
 
-def load_posts(request):
-    posts = Post.objects.all()
 
-    return JsonResponse([post.serialize() for post in posts], safe=False)
+def load_posts(request):
+    type = request.GET.get('type')
+
+    posts = Post.objects.none()
+
+    if type == 'all':
+        posts = Post.objects.all()
+
+    if type == 'user':
+        user_id = int(request.GET['user_id'])
+        profile_user = User(pk=user_id)
+
+        posts = Post.objects.filter(user=profile_user)
+        posts = posts.order_by("-timestamp").all()
+
+    if type == 'following_users':
+        following_users = Relationships.objects.filter(user_follow=request.user)
+
+        for f_user in following_users:
+            posts |= Post.objects.filter(user=f_user.user_followed)
+
+    try:
+        page = int(request.GET.get("page"))
+    except:
+        page = 1
+
+    posts = posts.order_by("-timestamp").all()
+
+    p = Paginator(posts, 10)
+    posts_page = p.page(page)
+
+    return JsonResponse({"posts":[post.serialize() for post in posts_page],
+    "has_next":posts_page.has_next(), "has_previous":posts_page.has_previous()}, safe=False)
 
 
 def login_view(request):
@@ -70,6 +101,7 @@ def register(request):
     else:
         return render(request, "network/register.html")
 
+
 @csrf_exempt 
 @login_required
 def new_post(request):
@@ -87,9 +119,6 @@ def load_profile(request):
     user_id = int(request.GET['user_id'])
     profile_user = User(pk=user_id)
 
-    posts = Post.objects.filter(user=profile_user)
-    posts = posts.order_by("-timestamp").all()
-
     try:
         is_following = Relationships.objects.get(user_follow=request.user, user_followed=profile_user)
         is_following = True
@@ -97,41 +126,22 @@ def load_profile(request):
         is_following = False
 
     return JsonResponse({"login_user_id": request.user.id,
-    "posts":[post.serialize() for post in posts], 
     "followers":Relationships.objects.filter(user_followed=profile_user).count(), 
     "following":Relationships.objects.filter(user_follow=profile_user).count(), 
     "is_following":is_following}, safe=False)
 
 
-def follow(request):
+def follow_unfollow(request):
+    type = request.GET.get('type')
+
     followed_id = int(request.GET['user_id'])
-    followed_user = User(pk = followed_id)
+    followed_user = User.objects.get(pk = followed_id)
 
-    # Follow relationship
-    following = Relationships(user_follow=request.user, user_followed=followed_user)
-    following.save()
+    if type == 'follow':
+        following = Relationships(user_follow=request.user, user_followed=followed_user)
+        following.save()
+        return JsonResponse({"message": f"{request.user} following {followed_user.username}."}, status=201)
 
-    return JsonResponse({"message": f"{request.user} following {followed_user.username}."}, status=201)
-
-
-def unfollow(request):
-    followed_id = int(request.GET['user_id'])
-    followed_user = User(pk = followed_id)
-
-    following = Relationships.objects.filter(user_follow=request.user, user_followed=followed_user)
+    following = Relationships.objects.get(user_follow=request.user, user_followed=followed_user)
     following.delete()
-
     return JsonResponse({"message": f"{request.user} unfollowing {followed_user.username}."}, status=201)
-
-
-@login_required
-def following_user_posts(request):
-    following_users = Relationships.objects.filter(user_follow=request.user)
-    
-    posts = Post.objects.none()
-    print(f'users: {following_users}')
-    for f_user in following_users:
-        posts |= Post.objects.filter(user=f_user.user_followed)
-
-    print(f'posts: {posts}')    
-    return JsonResponse([post.serialize() for post in posts], safe=False)
